+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.2] - 2023-12-22
+### Fixed
+- bug when selecting max score [#241](https://github.com/Qrlew/qrlew/pull/241)
+
 ## [0.7.1] - 2023-12-22
 ### Fixed
 - postrgres translation for some queries [#240](https://github.com/Qrlew/qrlew/pull/240)